+++ packages/stream_chat_flutter/lib/src/attachment/thumbnail/file_attachment_thumbnail.dart
@@ -4,6 +4,7 @@ import 'package:stream_chat_flutter/src/attachment/thumbnail/thumbnail_error.dar
 import 'package:stream_chat_flutter/src/attachment/thumbnail/video_attachment_thumbnail.dart';
 import 'package:stream_chat_flutter/src/utils/helpers.dart';
 import 'package:stream_chat_flutter_core/stream_chat_flutter_core.dart';
+import 'package:stream_core_flutter/stream_core_flutter.dart';
 
 /// {@template streamFileAttachmentThumbnail}
 /// Widget for building file attachment thumbnail.
@@ -58,7 +59,10 @@ class StreamFileAttachmentThumbnail extends StatelessWidget {
         fit: fit,
       ),
       // Return a generic file type icon.
-      _ => getFileTypeImage(mediaType?.mimeType),
+      _ => StreamFileTypeIcon.fromMimeType(
+        size: .xl,
+        mimeType: mediaType?.mimeType,
+      ),
     };
   }
 }