+++ packages/stream_chat_flutter/lib/src/components/message_composer/message_composer_input_header.dart
@@ -39,7 +39,7 @@ class _DefaultStreamMessageComposerInputHeader extends StatelessWidget {
   }
 
   Widget _buildContent(BuildContext context) {
-    final isEditing = !controller.message.state.isInitial;
+    final isEditing = controller.isEditing;
     final quotedMessage = !isEditing ? controller.message.quotedMessage : null;
     final ogAttachment = controller.ogAttachment;
     final nonOGAttachments = controller.attachments
@@ -77,7 +77,7 @@ class _DefaultStreamMessageComposerInputHeader extends StatelessWidget {
               Padding(
                 padding: contentPadding,
                 child: _EditMessageInHeader(
-                  message: controller.editingOriginalMessage ?? controller.message,
+                  message: controller.messageBeingEdited ?? controller.message,
                   onRemovePressed: controller.cancelEditMessage,
                 ),
               )