+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_ca.dart
@@ -49,9 +49,9 @@ class StreamChatLocalizationsCa extends GlobalStreamChatLocalizations {
 
   @override
   String attachmentsUploadProgressText({
-    required int remaining,
+    required int completed,
     required int total,
-  }) => 'Transferència en curs $remaining/$total ...';
+  }) => 'Pujats $completed de $total ...';
 
   @override
   String pinnedByUserText({